// src/enums/appEnum.ts
// 系统级别枚举

// 菜单类型
export enum MenuTypeEnum {
  LEFT = "left",
  TOP = "top",
  TOP_LEFT = "top-left",
}

// App theme enum
export enum SystemThemeEnum {
  DARK = "dark",
  LIGHT = "light",
  AUTO = "auto",
}

// Menu theme enum
export enum MenuThemeEnum {
  DARK = "dark",
  LIGHT = "light",
  DESIGN = "design",
}

// Menu open or close width
export enum MenuWidth {
  OPEN = "250px",
  CLOSE = "70px",
}

// Language
export enum LanguageEnum {
  ZH = "zh",
  EN = "en",
}